[JXD-254] Перенес все рендеры, в том числе и лямбду в display_render_base
