+++ esphome/const.py
@@ -862,6 +862,7 @@ CONF_SWING_MODE_STATE_TOPIC = "swing_mode_state_topic"
 CONF_SWING_OFF_ACTION = "swing_off_action"
 CONF_SWING_VERTICAL_ACTION = "swing_vertical_action"
 CONF_SWITCH_DATAPOINT = "switch_datapoint"
+CONF_SWITCH = "switch"
 CONF_SWITCHES = "switches"
 CONF_SYNC = "sync"
 CONF_TABLET = "tablet"
